Move out preferredApsRow
In order to break dependency cycles
